+++ lib/error_helper.h
@@ -27,80 +27,7 @@ limitations under the License.
 #include "lib/cstring.h"
 #include "lib/source_file.h"
 #include "lib/stringify.h"
-
-struct ErrorMessage {
-    enum class Type : std::size_t { None, Error, Warning };
-
-    Type type = Type::None;
-    std::string prefix = "";  // Typically error/warning type from catalog
-    std::string message = "";  // Actual formatted message
-    std::vector<Util::SourceInfo> locations = {}; // Relevant source locations for this error
-    std::string suffix = "";  // Used by errorWithSuffix
-
-/*
-    Util::SourceInfo position;  // Position in P4 source
-    std::string tail = "";  // Extra P4 source positions
-*/
-
-    ErrorMessage() {}
-    // Invoked from backwards compatible error_helper
-    ErrorMessage(const std::string &prefix, const Util::SourceInfo &info, std::string &suffix)
-        : prefix(prefix), locations({info}), suffix(suffix) {}
-    // Invoked from error_reporter
-    ErrorMessage(Type type, const std::string &prefix, const std::string &suffix)
-        : type(type), prefix(prefix), suffix(suffix) {}
-
-    std::string getPrefix() const {
-        std::string p = prefix;
-        if (type == Type::Error) {
-            if (p.empty()) p = "error: ";
-            else p = "[--Werror=" + p + "] error: ";
-        } else if (type == Type::Warning) {
-            if (p.empty()) p = "warning: ";
-            else p = "[--Wwarn=" + p + "] warning: ";
-        }
-        return p;
-    }
-
-    // original implementation, TODO: delete this
-    /*std::string _toString() const {
-        std::string result = position.toPositionString().c_str();
-        if (!result.empty()) result += ": ";
-        result += getPrefix() + message + "\n" + tail + suffix;
-        return result;
-    }*/
-
-    std::string toString() const {
-        std::string result = "";
-        std::string mainFragment = "";
-        if (!locations.empty() && locations.front().isValid()) {
-            result += locations.front().toPositionString() + ": ";
-            mainFragment = locations.front().toSourceFragment();
-        }
-
-        result += getPrefix() + message + "\n" + mainFragment;
-
-        for (unsigned i = 1; i < locations.size(); i++) {
-            if (locations[i].isValid())
-                result += locations[i].toPositionString() + "\n" + locations[i].toSourceFragment();
-        }
-
-        result += suffix;
-
-        return result;
-    }
-};
-
-struct ParserErrorMessage {
-    Util::SourceInfo location;
-    cstring message;
-
-    ParserErrorMessage(const Util::SourceInfo &loc, const cstring &msg) : location(loc), message(msg) {}
-
-    std::string toString() const {
-        return std::string(location.toPositionString().c_str()) + ":" + message + "\n" + location.toSourceFragment().c_str();
-    }
-};
+#include "lib/error_message.h"
 
 namespace priv {
 
@@ -209,16 +136,6 @@ auto error_helper(boost::format& f, ErrorMessage out, const T& t, Args... args) 
 template<class... Args>
 ErrorMessage error_helper(boost::format& f, ErrorMessage out, const Util::SourceInfo &info,
                          Args... args) {
-    /*cstring posString = info.toPositionString();
-    if (!out.position.isValid()) { // if there is not already something stored in out.position
-        out.position = info;
-        posString = "";
-    } else {
-        if (!posString.isNullOrEmpty())
-            posString += "\n";
-    }
-    out.tail += posString + info.toSourceFragment();*/
-
     if (info.isValid()) out.locations.push_back(info);
     return error_helper(f % "", out, std::forward<Args>(args)...);
 }
@@ -226,16 +143,6 @@ ErrorMessage error_helper(boost::format& f, ErrorMessage out, const Util::Source
 template<typename T, class... Args>
 auto error_helper(boost::format& f, ErrorMessage out, const T *t, Args... args) ->
     typename std::enable_if<std::is_base_of<Util::IHasSourceInfo, T>::value, ErrorMessage>::type {
-    /*cstring posString = t->getSourceInfo().toPositionString();
-    if (!out.position.isValid()) {
-        out.position = t->getSourceInfo();
-        posString = "";
-    } else {
-        if (!posString.isNullOrEmpty())
-            posString += "\n";
-    }
-    out.tail += posString + t->getSourceInfo().toSourceFragment();*/
-
     auto info = t->getSourceInfo();
     if (info.isValid()) out.locations.push_back(info);
     return error_helper(f % t->toString(), out, std::forward<Args>(args)...);
@@ -244,15 +151,6 @@ auto error_helper(boost::format& f, ErrorMessage out, const T *t, Args... args) 
 template<typename T, class... Args>
 auto error_helper(boost::format& f, ErrorMessage out, const T &t, Args... args) ->
     typename std::enable_if<std::is_base_of<Util::IHasSourceInfo, T>::value, ErrorMessage>::type {
-    /*cstring posString = t.getSourceInfo().toPositionString();
-    if (!out.position.isValid()) {
-        out.position = t.getSourceInfo();
-        posString = "";
-    } else {
-        if (!posString.isNullOrEmpty())
-            posString += "\n";
-    }
-    out.tail += posString + t.getSourceInfo().toSourceFragment();*/
 
     auto info = t.getSourceInfo();
     if (info.isValid()) out.locations.push_back(info);
